videorecorder: Fix off-by-one frame indices
This would only have been noticeable by comparing the embedded pts from
the video file with the tsync file data (the latter was incorrect).
Unfortunately we only noticed this now, and it can be (depending on how
the data is read) a pretty severe bug.

⚠️ If you were using `edlio`, you are *fine* with the old and the new
code. But if you were doing some manual mapping/parsing, you may need to
check your data.

+++ modules/videorecorder/encodehelper/videoreader.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2024 Matthias Klumpp <matthias@tenstral.net>
+ * Copyright (C) 2019-2026 Matthias Klumpp <matthias@tenstral.net>
  *
  * Licensed under the GNU Lesser General Public License Version 3
  *
@@ -49,7 +49,7 @@ public:
     ssize_t lastFrameIndex() const;
     double framerate() const;
 
-    std::optional<std::pair<cv::Mat, int64_t>> readFrame();
+    std::optional<std::pair<cv::Mat, size_t>> readFrame();
 
 private:
     class Private;
